+++ src/main/views/webContentEvents.ts
@@ -274,7 +274,7 @@ export class WebContentsEventManager {
                     // currently changing the userAgent for popup windows to allow plugins to go through google's oAuth
                     // should be removed once a proper oAuth2 implementation is setup.
                     popup.loadURL(details.url, {
-                        userAgent: composeUserAgent(),
+                        userAgent: composeUserAgent(false, serverURL),
                     });
                 }